requirements.txt
langchain
langchain-google-genai
streamlit
python-dotenv
pymongo